move featureToVector to util, so that other methods could use it to allow array input
